pyproject.toml
[project]
name = "ai-research-agent"
version = "0.1.0"
description = "AI Research Agent - Automatic research and report generation"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.28.0",
    "openai>=1.3.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "python-dotenv>=1.0.0",
    "sqlite3>=3.0",
    "markdown>=3.5.0",
    "reportlab>=4.0.0",
    "duckduckgo-search>=3.9.0",
    "loguru>=0.7.0",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
